refactor(app-manager): migrate ClassicyAppFinderOpenFile routing out of kernel (#006-d4c3)
Move the ClassicyAppFinderOpenFile cross-app orchestration logic from
the kernel router into the ClassicyAppFinder plugin (FinderContext.tsx).

- Export dispatchToPlugin from ClassicyAppManager so plugins can do
  cross-app dispatch without importing other app modules directly
- Remove MoviePlayerAppInfo import and ClassicyAppFinderOpenFile
  special-case from classicyDesktopStateEventReducer
- ClassicyAppFinder plugin now owns: QuickTime _creator routing,
  file-type handler lookup, and Finder fallback/error path
- Test files import FinderContext and MoviePlayerContext as side-effects
  to ensure plugins are registered before reducer tests run
